restructured front-end architecture
