+++ entity/init.sh
