--- presentation/front_end/homepage/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class HomepageConfig(AppConfig):
-    name = 'homepage'
