--- presentation/front_end/pet_detail/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class PetDetailConfig(AppConfig):
-    name = 'pet_detail'
