+++ presentation/front_end/templates/front_end/homepage.html
@@ -1,0 +1,5 @@
+<!DOCTYPE html>
+
+<html>
+    {{ results }}
+</html>
