+++ presentation/init.sh
